+++ Js2IL.Tests/IntrinsicCallables/GeneratorTests.cs
@@ -56,6 +56,12 @@ namespace Js2IL.Tests.IntrinsicCallables
         }
 
         [Fact]
+        public Task IntrinsicCallables_GlobalThis_Basic()
+        {
+            return GenerateTest(nameof(IntrinsicCallables_GlobalThis_Basic));
+        }
+
+        [Fact]
         public Task IntrinsicCallables_Symbol_Callable_Basic()
         {
             return GenerateTest(nameof(IntrinsicCallables_Symbol_Callable_Basic));